task: restructure file organization
