+++ src/main/java/com/kelechi/andela/hellobooks/service/UsersServiceImpl.java
@@ -12,12 +12,12 @@ import java.util.List;
 
 
 @Service
-public class SignupServiceImpl implements SignupService {
+public class UsersServiceImpl implements UsersService {
     @Autowired
     private UserRepository userRepository;
 
     @Override
-    public AuthenticationResponse signupUser(UserDTO userDTO) {
+    public AuthenticationResponse signup(UserDTO userDTO) {
         String username = userDTO.getUsername();
         String firstname = userDTO.getFirstname();
         String lastname = userDTO.getLastname();
@@ -93,5 +93,47 @@ public class SignupServiceImpl implements SignupService {
             }
         }
         return null;
+    }
+
+    @Override
+    public AuthenticationResponse signin(UserDTO userDTO) {
+
+        String email = userDTO.getEmail();
+        String password = userDTO.getPassword();
+
+        if("".equals(email) || email == null){
+            return new AuthenticationResponse(false, "Email is required");
+        }
+
+        if("".equals(password) || password == null){
+            return new AuthenticationResponse(false, "Password is required");
+        }
+
+        List<Users> existingUserList = userRepository.findByEmail(email);
+        if(existingUserList.size() == 0){
+            return new AuthenticationResponse(false, "Email or password is incorrect");
+        }else{
+            Users existingUser = existingUserList.get(0);
+            boolean isAuthenticated = checkPassword(password, existingUser.getPassword());
+            Long id = existingUser.getId();
+            String userEmail = existingUser.getEmail();
+            String username = existingUser.getUsername();
+
+            if(isAuthenticated){
+                String token = JSONTokenManager.createJWT(id, userEmail, username);
+                String message = "User successfully signed in";
+                boolean success = true;
+                return new AuthenticationResponse(token, success, message, existingUser);
+            }
+            return new AuthenticationResponse(false, "Email or password is incorrect");
+        }
+    }
+
+    private boolean checkPassword(String password, String hashedPassword){
+        boolean isVerified = false;
+        if(BCrypt.checkpw(password, hashedPassword)){
+            isVerified = true;
+        }
+        return  isVerified;
     }
 }
