--- src/main/java/com/kelechi/andela/hellobooks/service/authentication/SigninService.java
@@ -1,8 +1,0 @@
-package com.kelechi.andela.hellobooks.service.authentication;
-
-import com.kelechi.andela.hellobooks.dto.UserDTO;
-import com.kelechi.andela.hellobooks.util.response.AuthenticationResponse;
-
-public interface SigninService {
-    AuthenticationResponse signinUser(UserDTO userDTO);
-}
